requirements.txt
async-timeout==4.0.3
click==8.1.7
Flask==2.2.5
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
Pillow==9.5.0
redis==5.0.4
rq==1.16.2
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0